Move sockaddr_storage init function to libmptcpd.
The internal initialize_sockaddr_storage() function is now in
libmptcpd as mptcpd_sockaddr_storage_init() so that it may be used by
other function inside libmptcpd.
